Many improvements for the web UI. You can now use it to compile the test.

+++ test-1.2.sh
@@ -7,15 +7,19 @@ fi
 
 cd "$(dirname "$0")"
 
-[ ! -d build ] && mkdir build
+[ -d build ] && rm -r build
+mkdir build
 cd build
 echo "Compiling..."
+emcc --version | head -1 > emcc_version.txt
 
-for path in ../SDL-1.2.*/*.c; do
+for path in ../SDL-1.2.*/*.{c,cpp}; do
+    [ ! -f "$path" ] && continue
+    
     file="$(basename "$path")"
     [ -f "$file.js" ] && rm "$file.js"
     [ -f "$file.log" ] && rm "$file.log"
-    [ -f "$file.map" ] && rm "$file.map"
+    [ -f "$file.js.map" ] && rm "$file.js.map"
     
     echo "$file"
     emcc -g4 -O2 -Xclang -fcolor-diagnostics -o "$file.js" "$path" 2>&1 | tee "$file.log"
